Create README.md
In this topic, we learnt that how the computer system works.We learnt that how we communicate with the machine.How our language is different from their language.The computer system understands only the binary language which holds the coding system based on two numerical digits 0 and 1.Whatever the commands are given to the computer.They take time to process those input data through converting those data into binary language and arrange into the proper places according to its preferable arrangement and interpret it to understand.And then forms its own answer and then gives the result as output.The input data arranged in form of bits(again using the digits 0 and 1 to understand).The computer system perform arithmetically as well as logical operations.Arithmetic operations include the addition, subtraction, multiplication and division.The logical operations include AND, OR,NOT which helps them to understand the commands to do their operations.We also learnt the 1’s complement, 2’s complement  and conversion of binary language into decimal form and vice versa.We do exponential power of the binary language and with respect to their palace value and add them to get the decimal form of number. And to get the binary number we keep on dividing the number by 2.This is how we get the binary number.Through this we learnt that how the computer understand the language and the number of digits formed in binary number called the bits.This is the very basic concept of the computer language that helps us in communicating with the machines.